+++ sentry/src/main/java/io/sentry/util/ExceptionUtils.java
@@ -1,0 +1,23 @@
+package io.sentry.util;
+
+import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.NotNull;
+
+@ApiStatus.Internal
+public final class ExceptionUtils {
+
+  /**
+   * Returns exception root cause or the exception itself if there are no causes
+   *
+   * @param throwable - the throwable
+   * @return the root cause
+   */
+  public static @NotNull Throwable findRootCause(final @NotNull Throwable throwable) {
+    Objects.requireNonNull(throwable, "throwable cannot be null");
+    Throwable rootCause = throwable;
+    while (rootCause.getCause() != null && rootCause.getCause() != rootCause) {
+      rootCause = rootCause.getCause();
+    }
+    return rootCause;
+  }
+}