+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Unreleased
 
+* Fix: associate event with transaction when thrown exception is not a direct cause (#1463)
+
 # 5.0.0-beta.2
 
 * Fix: sentry-android-timber package sets sentry.java.android.timber as SDK name (#1456)
 * Fix: When AppLifecycleIntegration is closed, it should remove observer using UI thread (#1459)
 * Bump: AGP to 4.2.0 (#1460)
 * Fix: NPE when adding Context Data with null values for log4j2 (#1465)
-
 
 Breaking Changes: